feat: add pop! and update functions for NMEAData, move to parse.jl
